flask_app/requirements.txt
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
xmltodict==0.12.0
